requirements.txt
fastapi==0.101.0
sqlalchemy==2.0.19
uvicorn==0.23.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
psycopg2-binary==2.9.7
httpagentparser==1.9.5
python-multipart==0.0.6
requests==2.31.0
pytz==2023.3
pydantic-settings==2.1.0
inflect ==7.0.0
sqlalchemy_utils ==0.41.1
selenium==4.23.1
elasticsearch==8.11.1